bridge.sol: extract message related fns into `Message` library
